move webpack_dev_server config to separate file #25
  webpack dev server can now be loaded via dev_server.js
  the inline mode for hmr isnt present in webpack.config but is rather
  appended with "unshift" in the dev_server config.
  Also removed some options from url-loader loader.
