+++ webpack.config.js
@@ -1,38 +1,30 @@
 var path = require('path');
 var webpack = require('webpack');
-var webpackDevServer=require('webpack-dev-server');
-var port = 8090;
 var source_dir = path.resolve(__dirname, 'app/Resources/AppBundle/assets');
-var dest_dir = path.resolve(__dirname, 'web/assets');
+var dest_dir = path.resolve(__dirname, 'web/assets/');
 var bower = path.resolve(__dirname,'vendor/bower_components');
 var node = path.resolve(__dirname,'node_modules');
 
 var config = {
     entry: {
         jquery: [
-            'webpack-dev-server/client?http://localhost:8090',
-            'webpack/hot/dev-server',
-            path.resolve(bower, "jquery/dist/jquery.js"),
+            path.resolve(bower, "jquery/dist/jquery.js")
         ],
         vendor: [
-            'webpack-dev-server/client?http://localhost:8090',
-            'webpack/hot/dev-server',
-            // path.resolve(bower, "bootstrap/less/bootstrap.less"),
+            // // path.resolve(bower, "bootstrap/less/bootstrap.less"),
             // path.resolve(bower, "bootstrap/dist/js/bootstrap.js"),
             path.resolve(bower, "Materialize/sass/materialize.scss"),
-            path.resolve(bower, "Materialize/dist/js/materialize.js"),
+            path.resolve(bower, "Materialize/dist/js/materialize.js")
         ],
         main: [
-            'webpack-dev-server/client?http://localhost:8090',
-            'webpack/hot/dev-server',
-            path.resolve(source_dir, "js/hello.js"),
-            path.resolve(source_dir, "css/hello.scss"),
+            path.resolve(source_dir, "js/main.js"),
+            path.resolve(source_dir, "css/main.scss")
         ]
     },
     output: {
         filename: '[name].js',
-        path: dest_dir,
-        publicPath: "http://localhost:8090/assets/"
+        path: dest_dir + "/",
+        publicPath:  dest_dir + "/" //"http://localhost:8090/assets/"
     },
     devtool: 'source-maps',
     module: {
@@ -52,7 +44,7 @@ var config = {
             },
             {
                 test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-                loader: "url-loader?limit=10000&mimetype=application/font-woff"
+                loader: "url-loader"
             },
             {
                 test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/,
@@ -66,7 +58,7 @@ var config = {
         alias: {
             jquery: "vendor/bower_components/jquery/src/jquery.js",
             hammerjs: "vendor/bower_components/Materialize/js/hammer.min.js"
-        },
+        }
     },
     resolveLoader: {
         root: [node, bower]
@@ -75,32 +67,12 @@ var config = {
         new webpack.HotModuleReplacementPlugin(),
         new webpack.ResolverPlugin(
             new webpack.ResolverPlugin.DirectoryDescriptionFilePlugin(".bower.json", ["main"])
-        ),
+        )
         // new webpack.ProvidePlugin({
         //     $: "jquery",
         //     jQuery: "jquery",
         // })
     ]
 };
-
-var devServer = new webpackDevServer(webpack(config), {
-    hot: true,
-    publicPath: "http://localhost:8090/assets/",
-    contentBase: "web/",
-    headers: {"X-Custom-Header": "yes"},
-    stats: {colors: true}
-    }
-);
-
-devServer.listen(
-    8090,
-    'localhost',
-    function (err, result) {
-        if (err) {
-            console.log(err);
-        }
-        console.log('Listening at localhost:'+port);
-    }
-);
 
 module.exports = config;
